Tighten up schema types
