+++ discoveryapis_generator/lib/src/dart_schema_types.dart
@@ -6,83 +6,11 @@
 
 import 'dart_api_library.dart';
 import 'dart_comments.dart';
+import 'dart_schema_type.dart';
 import 'dart_schemas.dart';
 import 'json_type.dart';
 import 'namer.dart';
 import 'utils.dart';
-
-/// Represents an internal representation used for codegen.
-///
-/// [DartSchemaType] and it's subclasses are a representation for codegen of:
-///   - dart class definitions
-///   - dart type declarations
-///   - dart expressions for encoding/decoding json
-///
-/// Before a [DartSchemaType] can be used, it's [resolve] method must be called
-/// to resolve all forward references.
-abstract class DartSchemaType {
-  // [className] is the name of the dart class this [DartSchemaType] represents
-  // or `null` if it does not represent a schema type represented by a custom
-  // dart class.
-  final Identifier? className;
-  final Comment comment;
-  final DartApiImports imports;
-
-  bool _resolved = false;
-
-  DartSchemaType(this.imports, this.className, {Comment? comment})
-      : comment = comment ?? Comment.empty;
-
-  DartSchemaType? resolve(DartSchemaTypeDB db) {
-    if (!_resolved) {
-      _resolved = true;
-      return _resolve(db);
-    }
-    return this;
-  }
-
-  DartSchemaType? _resolve(DartSchemaTypeDB db);
-
-  String? get declaration;
-
-  JsonType get jsonType;
-
-  /// [value] is the string expression of this [DartSchemaType] that needs to be
-  /// encoded.
-  ///
-  /// This method is used for encoding parameter types for the URI query part.
-  String? primitiveEncoding(String? value);
-
-  /// Whether this value needs a primitive encoding.
-  bool get needsPrimitiveEncoding => primitiveEncoding('foo') != 'foo';
-
-  /// [value] is the string expression of this [DartSchemaType] that needs to be
-  /// encoded.
-  String jsonEncode(String value);
-
-  /// [json] is the string expression of json data that needs to be decoded to
-  /// a [DartSchemaType].
-  String jsonDecode(String json);
-
-  /// Whether this value needs a JSON encoding or not.
-  bool get needsJsonEncoding => jsonEncode('foo') != 'foo';
-
-  /// Whether this value needs a JSON decoding or not.
-  bool get needsJsonDecoding => jsonDecode('foo') != 'foo';
-
-  String get coreMapJsonTypeArguments =>
-      '<${imports.core.ref()}String, ${imports.core.ref()}dynamic>';
-
-  String get coreMapJsonType =>
-      '${imports.core.ref()}Map$coreMapJsonTypeArguments';
-
-  String decodeFromMap(String jsonName) {
-    final decodeString = jsonDecode("_json['${escapeString(jsonName)}']");
-
-    return '_json.containsKey'
-        "('${escapeString(jsonName)}') ? $decodeString : null";
-  }
-}
 
 /// Class representing non-primitive types.
 ///
@@ -125,7 +53,7 @@ class ObjectType extends ComplexDartSchemaType {
         super(imports, name, comment: comment);
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) {
+  DartSchemaType resolveCore(DartSchemaTypeDB db) {
     for (var i = 0; i < properties.length; i++) {
       final property = properties[i];
       final resolvedProperty = DartClassProperty(property.name,
@@ -266,7 +194,7 @@ class AbstractVariantType extends ComplexDartSchemaType {
         super(imports, name, comment: comment);
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) {
+  DartSchemaType resolveCore(DartSchemaTypeDB db) {
     map.forEach((String? name, DartSchemaType ref) {
       final resolvedType = ref.resolve(db);
       if (resolvedType is ObjectType) {
@@ -348,7 +276,7 @@ class DartSchemaForwardRef extends DartSchemaType {
   }
 
   @override
-  DartSchemaType? _resolve(DartSchemaTypeDB db) => null;
+  DartSchemaType? resolveCore(DartSchemaTypeDB db) => null;
 
   @override
   JsonType get jsonType {
@@ -387,7 +315,7 @@ abstract class PrimitiveDartSchemaType extends DartSchemaType {
   PrimitiveDartSchemaType(DartApiImports imports) : super(imports, null);
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) => this;
+  DartSchemaType resolveCore(DartSchemaTypeDB db) => this;
 
   @override
   String? primitiveEncoding(String? value) => "'\${$value}'";
@@ -583,7 +511,7 @@ class UnnamedArrayType extends ComplexDartSchemaType implements HasInnertype {
       : super(imports, null);
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) {
+  DartSchemaType resolveCore(DartSchemaTypeDB db) {
     innerType = innerType!.resolve(db);
     return this;
   }
@@ -636,7 +564,7 @@ class NamedArrayType extends ComplexDartSchemaType implements HasInnertype {
       : super(imports, name, comment: comment);
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) {
+  DartSchemaType resolveCore(DartSchemaTypeDB db) {
     innerType = innerType!.resolve(db);
     return this;
   }
@@ -728,7 +656,7 @@ class UnnamedMapType extends ComplexDartSchemaType {
   }
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) {
+  DartSchemaType resolveCore(DartSchemaTypeDB db) {
     fromType = fromType!.resolve(db);
     toType = toType!.resolve(db);
     return this;
@@ -796,7 +724,7 @@ class NamedMapType extends ComplexDartSchemaType {
   }
 
   @override
-  DartSchemaType _resolve(DartSchemaTypeDB db) {
+  DartSchemaType resolveCore(DartSchemaTypeDB db) {
     fromType = fromType!.resolve(db);
     toType = toType!.resolve(db);
     return this;
