+++ discoveryapis_generator/test/dart_schemas_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:discoveryapis_generator/src/dart_api_library.dart';
+import 'package:discoveryapis_generator/src/dart_schema_type.dart';
 import 'package:discoveryapis_generator/src/dart_schema_types.dart';
 import 'package:discoveryapis_generator/src/dart_schemas.dart';
 import 'package:discoveryapis_generator/src'
